[script.image.resource.select] 3.0.2
