+++ script.image.resource.select/changelog.txt
@@ -1,3 +1,6 @@
+v3.0.2
+- refactor
+
 v3.0.1
 - bump
